+++ src/iotsaBLEServer.cpp
@@ -118,13 +118,6 @@ void IotsaBLEServerMod::createServer() {
   iotsaConfig.ensureConfigLoaded();
   IFBLEDEBUG IotsaSerial.print("BLE hostname: ");
   IFBLEDEBUG IotsaSerial.println(iotsaConfig.hostName.c_str());
-  #ifndef IOTSA_WITH_NIMBLE
-  // We de-init bluetooth and release classic-mode memory, in the hope
-  // this frees some of the memory the bluedroid stack uses.
-  BLEDevice::deinit(false);
-  esp_bt_controller_mem_release(ESP_BT_MODE_CLASSIC_BT);
-  #endif
-
   iotsaBLE_ensureInitialized();
   BLEDevice::setMTU(BLE_ATT_MTU_MAX);
   if (tx_power >= 0) {
@@ -369,20 +362,11 @@ void IotsaBleApiService::addCharacteristic(UUIDstring charUUID, int mask, uint8_
   }
   BLECharacteristic *newChar = bleService->createCharacteristic(charUUID, mask);
   newChar->setCallbacks(new IotsaBLECharacteristicCallbacks(charUUID, apiProvider));
-#ifdef IOTSA_WITH_NIMBLE
   BLEDescriptor *d2901 = newChar->createDescriptor("2901");
   BLE2904 *d2904 = newChar->create2904();
-#else
-  BLEDescriptor *d2901 = new BLEDescriptor("2901");
-  BLE2904 *d2904 = new BLE2904();
-#endif
   d2901->setValue(std::string(d2901descr));
   d2904->setFormat(d2904format);
   d2904->setUnit(d2904unit);
-#ifndef IOTSA_WITH_NIMBLE
-  newChar->addDescriptor(d2901);
-  newChar->addDescriptor(d2904);
-#endif
 
   characteristicUUIDs[nCharacteristic-1] = charUUID;
   bleCharacteristics[nCharacteristic-1] = newChar;